refactor: only expose the data needed from API request
Returns only the proof request ID after the API call to create the request. This makes it explicit that this is the only field expected to be used.

I'm a bit uncertain that it should be the responsibility of this function call to choose what fields could be used. However, it seems that this is better than exposing data without providing a schema for it.